Draft split of `CompactionService` into planner and execution parts
