+++ src/domain/core/src/services/compaction/mod.rs
@@ -1,0 +1,16 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+mod compaction_execution_service;
+mod compaction_listener;
+mod compaction_planner;
+
+pub use compaction_execution_service::*;
+pub use compaction_listener::*;
+pub use compaction_planner::*;
