+++ src/domain/task-system/domain/src/services/task_definition_planner.rs
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0.
 
 use internal_error::InternalError;
-use kamu_core::{CompactionOptions, PullOptions, PullPlanIterationJob, ResolvedDataset};
+use kamu_core::{CompactionPlan, PullOptions, PullPlanIterationJob, ResolvedDataset};
 use opendatafabric::Multihash;
 
 use crate::{LogicalPlan, LogicalPlanProbe};
@@ -62,7 +62,7 @@ pub struct TaskDefinitionReset {
 #[derive(Debug)]
 pub struct TaskDefinitionHardCompact {
     pub target: ResolvedDataset,
-    pub compaction_options: CompactionOptions,
+    pub compaction_plan: CompactionPlan,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
