+++ src/domain/task-system/services/src/task_definition_planner_impl.rs
@@ -22,6 +22,7 @@ pub struct TaskDefinitionPlannerImpl {
     dataset_registry: Arc<dyn DatasetRegistry>,
     dataset_env_vars_svc: Arc<dyn DatasetEnvVarService>,
     pull_request_planner: Arc<dyn PullRequestPlanner>,
+    compaction_planner: Arc<dyn CompactionPlanner>,
     tenancy_config: Arc<TenancyConfig>,
 }
 
@@ -34,12 +35,14 @@ impl TaskDefinitionPlannerImpl {
         dataset_registry: Arc<dyn DatasetRegistry>,
         dataset_env_vars_svc: Arc<dyn DatasetEnvVarService>,
         pull_request_planner: Arc<dyn PullRequestPlanner>,
+        compaction_planner: Arc<dyn CompactionPlanner>,
         tenancy_config: Arc<TenancyConfig>,
     ) -> Self {
         Self {
             dataset_registry,
             dataset_env_vars_svc,
             pull_request_planner,
+            compaction_planner,
             tenancy_config,
         }
     }
@@ -139,9 +142,15 @@ impl TaskDefinitionPlannerImpl {
             keep_metadata_only: args.keep_metadata_only,
         };
 
+        let compaction_plan = self
+            .compaction_planner
+            .build_compaction_plan(target.clone(), compaction_options, None)
+            .await
+            .int_err()?;
+
         Ok(TaskDefinition::HardCompact(TaskDefinitionHardCompact {
             target,
-            compaction_options,
+            compaction_plan,
         }))
     }
 }
