+++ src/infra/core/src/compaction/mod.rs
@@ -1,0 +1,14 @@
+// Copyright Kamu Data, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+mod compaction_executor_impl;
+mod compaction_planner_impl;
+
+pub use compaction_executor_impl::*;
+pub use compaction_planner_impl::*;
